Move stuff into CutUtils.
Just want cut definitions and cutting mechanics in the Cuts.* files.
