+++ includes/Constants.h
@@ -10,7 +10,6 @@ class CVUniverse;
 //==============================================================================
 // Typedefs, enums, colors, constants
 //==============================================================================
-
 enum ECuts {
   kNoCuts,
   kGoodObjects,
